+++ providers/linux/host_fqdn_integration_linux_test.go
@@ -35,62 +35,6 @@ import (
 const wantHostname = "debian"
 const wantDomainCgo = "cgo"
 
-func TestHost_FQDN_Domain_Cgo(t *testing.T) {
-	host, err := newLinuxSystem("").Host()
-	if err != nil {
-		t.Fatal(fmt.Errorf("could not het host information: %w", err))
-	}
-
-	got := host.Info()
-	if got.Hostname != wantHostname {
-		t.Errorf("got wrong hostname want: %q, got %q", wantHostname, got.Hostname)
-	}
-	if got.Domain != wantDomainCgo {
-		t.Errorf("got wrong domain want: %q, got %q", wantDomainCgo, got.Domain)
-	}
-	if got.FQDN != fmt.Sprintf("%s.%s", wantHostname, wantDomainCgo) {
-		t.Errorf("FQDN shpould not be empty")
-	}
-}
-
-func TestHost_FQDN_No_Domain_Cgo(t *testing.T) {
-	host, err := newLinuxSystem("").Host()
-	if err != nil {
-		t.Fatal(fmt.Errorf("could not het host information: %w", err))
-	}
-
-	got := host.Info()
-	if got.Hostname != wantHostname {
-		t.Errorf("got wrong hostname want: %s, got %s", wantHostname, got.Hostname)
-	}
-	if got.Domain != "" {
-		t.Errorf("got wrong domain should be empty but got %s", got.Domain)
-	}
-	wantFQDN := fmt.Sprintf("%s.%s", wantHostname, "lan")
-	if got.FQDN != wantFQDN {
-		t.Errorf("got wrong FQDN, want: %s, got %s", wantFQDN, got.FQDN)
-	}
-}
-
-func TestHost_FQDN_Domain_NoCgo(t *testing.T) {
-	t.SkipNow()
-	host, err := newLinuxSystem("").Host()
-	if err != nil {
-		t.Fatal(fmt.Errorf("could not het host information: %w", err))
-	}
-
-	got := host.Info()
-	if got.Hostname != wantHostname {
-		t.Errorf("hostname want: %s, got %s", wantHostname, got.Hostname)
-	}
-	if got.Domain != "" {
-		t.Errorf("domain should be empty but got %s", got.Domain)
-	}
-	if got.FQDN != "" {
-		t.Errorf("FQDN should empty, got: %s", got.FQDN)
-	}
-}
-
 func TestHost_FQDN(t *testing.T) {
 	tcs := []struct {
 		name string
@@ -105,8 +49,11 @@ func TestHost_FQDN(t *testing.T) {
 				AttachStdout: testing.Verbose(),
 				WorkingDir:   "/usr/src/elastic/go-sysinfo",
 				Image:        "golang:1.19-bullseye",
-				Cmd: []string{"go", "test", "-v", "-run",
-					"^TestHost_FQDN_Domain_Cgo", "./providers/linux"},
+				Cmd: []string{
+					"go", "test", "-v",
+					"-tags", "integration,docker",
+					"-run", "^TestHost_FQDN_Domain_Cgo",
+					"./providers/linux"},
 				Tty: false,
 			},
 		},
@@ -118,8 +65,11 @@ func TestHost_FQDN(t *testing.T) {
 				AttachStdout: testing.Verbose(),
 				WorkingDir:   "/usr/src/elastic/go-sysinfo",
 				Image:        "golang:1.19-bullseye",
-				Cmd: []string{"go", "test", "-v", "-run",
-					"^TestHost_FQDN_No_Domain_Cgo", "./providers/linux"},
+				Cmd: []string{
+					"go", "test", "-v", "-count", "1",
+					"-tags", "integration,docker",
+					"-run", "^TestHost_FQDN_No_Domain_Cgo",
+					"./providers/linux"},
 				Tty: false,
 			},
 		},
@@ -132,8 +82,11 @@ func TestHost_FQDN(t *testing.T) {
 				Env:          []string{"CGO_ENABLED=0"},
 				WorkingDir:   "/usr/src/elastic/go-sysinfo",
 				Image:        "golang:1.19-bullseye",
-				Cmd: []string{"go", "test", "-v", "-run",
-					"^TestHost_FQDN_Domain_NoCgo", "./providers/linux"},
+				Cmd: []string{
+					"go", "test", "-v",
+					"-tags", "integration,docker",
+					"-run", "^TestHost_FQDN_Domain_NoCgo",
+					"./providers/linux"},
 				Tty: false,
 			},
 		},
@@ -141,7 +94,7 @@ func TestHost_FQDN(t *testing.T) {
 
 	cli, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
 	if err != nil {
-		panic(err)
+		t.Fatalf("failed to create docker client: %v", err)
 	}
 	defer cli.Close()
 
@@ -157,13 +110,13 @@ func runOnDocker(t *testing.T, cli *client.Client, cf *container.Config) {
 
 	pwd, err := os.Getwd()
 	if err != nil {
-		panic(err)
+		t.Fatalf("could not get current directory: %v", err)
 	}
 	wd := pwd + "../../../"
 
 	reader, err := cli.ImagePull(ctx, cf.Image, types.ImagePullOptions{})
 	if err != nil {
-		panic(err)
+		t.Fatalf("failed to pull image %s: %v", cf.Image, err)
 	}
 	defer reader.Close()
 	io.Copy(os.Stdout, reader)
@@ -173,40 +126,44 @@ func runOnDocker(t *testing.T, cli *client.Client, cf *container.Config) {
 		Binds:      []string{wd + ":/usr/src/elastic/go-sysinfo"},
 	}, nil, nil, "")
 	if err != nil {
-		panic(err)
+		t.Fatalf("could not create docker conteiner: %v", err)
 	}
 	defer func() {
 		err = cli.ContainerRemove(ctx, resp.ID, types.ContainerRemoveOptions{
 			Force: true, RemoveVolumes: true})
 		if err != nil {
-			panic(err)
+			t.Logf("WARNING: could not remove docker container: %v", err)
 		}
 	}()
 
 	if err := cli.ContainerStart(ctx, resp.ID, types.ContainerStartOptions{}); err != nil {
-		panic(err)
+		t.Fatalf("could not start docker container: %v", err)
 	}
 
 	statusCh, errCh := cli.ContainerWait(ctx, resp.ID, container.WaitConditionNotRunning)
 	select {
 	case err := <-errCh:
 		if err != nil {
-			panic(err)
+			// Not using fatal as we might be able to recover the container
+			// logs.
+			t.Errorf("docker ContainerWait failed: %v", err)
 		}
 	case s := <-statusCh:
 		if s.StatusCode != 0 {
-			var err error
+			msg := fmt.Sprintf("container exited with status code %d", s.StatusCode)
 			if s.Error != nil {
-				err = fmt.Errorf("container errored: %s", s.Error.Message)
+				msg = fmt.Sprintf("%s: error: %s", msg, s.Error.Message)
 			}
-			t.Errorf("conteiner exited with code %d: error: %v", s.StatusCode, err)
+
+			// Not using fatal as we might be able to recover the container
+			// logs.
+			t.Errorf(msg)
 		}
-		t.Log("docker starts channel:", s)
 	}
 
 	out, err := cli.ContainerLogs(ctx, resp.ID, types.ContainerLogsOptions{ShowStderr: true, ShowStdout: true})
 	if err != nil {
-		panic(err)
+		t.Fatalf("could not get container logs: %v", err)
 	}
 
 	stdcopy.StdCopy(os.Stdout, os.Stderr, out)
